Dynamically import ethereal
