+++ packages/engine/src/scene/behaviors/createPortal.ts
@@ -10,7 +10,7 @@ import { ColliderComponent } from '../../physics/components/ColliderComponent';
 import { CollisionGroups } from '../../physics/enums/CollisionGroups';
 import { PhysicsSystem } from '../../physics/systems/PhysicsSystem';
 import { TransformComponent } from '../../transform/components/TransformComponent';
-import { FontManager } from '../../ui/FontManager';
+import { FontManager } from '../../ui/classes/FontManager';
 import { Object3DComponent } from '../components/Object3DComponent';
 import { PortalComponent } from '../components/PortalComponent';
